# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciar pacientes, consultas e usuários do Hospital CESUPA.",
        "title": "CESUPA Hospital API",
        "termsOfService": "http://cesupa-hospital/terms/",
        "contact": {
            "name": "Equipe de Desenvolvimento",
            "url": "http://cesupa-hospital/support",
            "email": "suporte@cesupa-hospital.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pacients": {
            "get": {
                "description": "Retorna todos os pacientes, podendo filtrar por nome e/ou idade",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Lista pacientes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filtra pelo nome (substring)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtra pela idade exata",
                        "name": "age",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Pacient"
                            }
                        }
                    },
                    "404": {
                        "description": "No pacient was found",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Registra um paciente com dados obrigatórios e opcionais",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Cadastra um novo paciente",
                "parameters": [
                    {
                        "description": "Dados do paciente",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pacients.AddPacientDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Pacient"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create pacient",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pacients/{id}": {
            "get": {
                "description": "Retorna os dados de um paciente pelo seu ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Busca paciente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pacient"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pacient not found",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza campos de um paciente existente (PATCH semantics)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Atualiza paciente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados que serão atualizados",
                        "name": "paciente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pacients.UpdatePacientDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pacient"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or Input",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pacient not found",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update pacient",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Inativa (ou exclui) um paciente pelo ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Exclui paciente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Pacient"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pacient not found",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete pacient",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pacients/{id}/appointments": {
            "post": {
                "description": "Cria uma nova consulta para o paciente informado",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pacientes"
                ],
                "summary": "Agenda consulta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID do paciente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados da consulta",
                        "name": "appointment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pacients.ScheduleAppointmentDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Appointment"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or Input",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pacient not found",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create appointment",
                        "schema": {
                            "$ref": "#/definitions/pacients.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "enums.BloodType": {
            "type": "string",
            "enum": [
                "A+",
                "A-",
                "B+",
                "B-",
                "AB+",
                "AB-",
                "O+",
                "O-"
            ],
            "x-enum-varnames": [
                "APositive",
                "ANegative",
                "BPositive",
                "BNegative",
                "ABPositive",
                "ABNegative",
                "OPositive",
                "ONegative"
            ]
        },
        "enums.Role": {
            "type": "string",
            "enum": [
                "recepcionist",
                "doctor",
                "admin"
            ],
            "x-enum-varnames": [
                "Receptionist",
                "Doctor",
                "Admin"
            ]
        },
        "enums.Sex": {
            "type": "string",
            "enum": [
                "male",
                "famale"
            ],
            "x-enum-varnames": [
                "Male",
                "Female"
            ]
        },
...
        "models.Appointment": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "pacientId": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "models.Pacient": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "allergies": {
                    "type": "string"
                },
                "birthDate": {
                    "type": "string"
                },
                "bloodType": {
                    "$ref": "#/definitions/enums.BloodType"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "sex": {
                    "$ref": "#/definitions/enums.Sex"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "appointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Appointment"
                    }
                },
                "cpf": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/enums.Role"
                }
            }
        },
        "pacients.AddPacientDTO": {
            "type": "object",
            "required": [
                "address",
                "birthDate",
                "cpf",
                "name",
                "phoneNumber",
                "sex"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "allergies": {
                    "type": "string"
                },
                "birthDate": {
                    "type": "string"
                },
                "bloodType": {
                    "$ref": "#/definitions/enums.BloodType"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "sex": {
                    "$ref": "#/definitions/enums.Sex"
                }
            }
        },
        "pacients.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "pacients.ScheduleAppointmentDTO": {
            "type": "object",
            "required": [
                "date",
                "doctorId"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "doctorId": {
                    "type": "integer"
                }
            }
        },
        "pacients.UpdatePacientDTO": {
            "type": "object",
            "required": [
                "address",
                "birthDate",
                "cpf",
                "name",
                "phoneNumber",
                "sex"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "allergies": {
                    "type": "string"
                },
                "birthDate": {
                    "type": "string"
                },
                "bloodType": {
                    "$ref": "#/definitions/enums.BloodType"
                },
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "sex": {
                    "$ref": "#/definitions/enums.Sex"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}